src/backend/src/requirements.txt
asarPy==1.0.1
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels_redis==4.2.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.1
django-cors-headers==4.4.0
django-countries==7.6.1
django-filter==24.2
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Markdown==3.6
oauthlib==3.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-decouple==3.8
python3-openid==3.2.0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.1
zope.interface==6.4.post2
bleach==5.0.1
django-bleach==3.1.0
svgwrite==1.4.1  # Added svgwrite for generating SVG avatars
django-model-utils==4.4.0  # Added for FieldTracker to track field changes